fix: display steps with applied "margin-top" and/or "margin-bottom" correctly
- those steps were not placed correctly on vertical Y axis
- also added part="step-label" to the step label, so it's easier to style it using CSS
- also removed some of the console.log() calls, to reduce clutter in the javascript console
