+++ src/util/ganttUtil.ts
@@ -1,3 +1,5 @@
+import * as ElementUtil from 'tltv-timeline-element/dist/src/util/elementUtil.js';
+
 /**
  * Return x and y coordinates from the given event. Supports mouse and touch event.
  * @param event target Event
@@ -22,4 +24,23 @@ export function getPageX(event: Event, container: HTMLDivElement) {
 
 export function getPageY(event: Event, container: HTMLDivElement) {
     return getPointForEvent(event, container)[1];
+}
+
+export function getElementHeightWithMargin(div: HTMLElement): number {
+    let height: number = Math.round(ElementUtil.getHeight(div));
+    let marginHeight: number = 0;
+    marginHeight = getMarginByComputedStyle(div);
+    return height + Math.round(marginHeight);
+}
+
+function getMarginByComputedStyle(elem: HTMLElement): number {
+    let cs = elem.ownerDocument.defaultView.getComputedStyle(elem);
+    let size: number;
+    if (cs) {
+        size = parseInt(cs.getPropertyValue('margin-top'))
+            + parseInt(cs.getPropertyValue('margin-bottom'));
+    } else {
+        size = 0;
+    }
+    return size;
 }
